refactor(artifact-provenance): extract unindexed recovery
