+++ src/main/artifacts/provenance-repository.ts
@@ -1,5 +1,5 @@
 import { randomUUID } from 'node:crypto'
-import { mkdir, readFile, readdir, rename, rm, stat, writeFile } from 'node:fs/promises'
+import { readFile, rename, rm, stat, writeFile } from 'node:fs/promises'
 import { dirname, join } from 'node:path'
 import { pathToFileURL } from 'node:url'
 
@@ -25,7 +25,7 @@ import {
   MAX_ARTIFACT_VERSION_DESCRIPTOR_IDS,
   type ResolveArtifactVersionDescriptorsRequest
 } from '../../shared/artifacts'
-import { ArtifactCompatibilityScanIncompleteError, ArtifactRepository } from './repository'
+import { ArtifactRepository } from './repository'
 import { defaultArtifactDurability, type ArtifactDurability } from './durability'
 import {
   ArtifactProvenanceVersionWriter,
@@ -34,7 +34,7 @@ import {
 } from './provenance-version-writer'
 import { NotebookRunRepository } from '../notebook/repository'
 import type { NotebookRunInputFile } from '../../shared/notebook'
-import { canonicalJson, sha256, type CanonicalJson } from './provenance-canonical'
+import { canonicalJson, sha256 } from './provenance-canonical'
 import { ArtifactProvenanceProducerCapture } from './provenance-producer-capture'
 import {
   parseArtifactExecutionSnapshot,
@@ -51,6 +51,7 @@ import {
   type ArtifactProjectReconciliationSnapshot
 } from './provenance-finalization-recovery'
 import { ArtifactProvenanceStagingRecovery } from './provenance-staging-recovery'
+import { ArtifactProvenanceUnindexedRecovery } from './provenance-unindexed-recovery'
 import { readOptionalFile, resolveStorageKey, storageKey } from './provenance-storage'
 import { toCheck, toReview } from '../reviewer/repository'
 import { selectReviewChainForArtifactVersion } from '../reviewer/artifact-version-review'
@@ -65,7 +66,6 @@ import type { PersistedChatSession } from '../../shared/session-persistence'
 import { sanitizeActivityGroup, sanitizeToolActivity } from '../../shared/session-persistence'
 
 const SAFE_SEGMENT_PATTERN = /^[A-Za-z0-9][A-Za-z0-9._-]*$/
-const SHA256_PATTERN = /^[a-f0-9]{64}$/
 
 type ArtifactProvenanceRepositoryOptions = {
   storageRoot: string
@@ -125,12 +125,6 @@ const hasServerInferredProducer = (evidenceJson: string): boolean => {
     return false
   }
 }
-
-const stringValue = (value: unknown): string | undefined =>
-  typeof value === 'string' ? value : undefined
-
-const numberValue = (value: unknown): number | undefined =>
-  typeof value === 'number' && Number.isFinite(value) ? value : undefined
 
 type PersistedExecutionInputRow = {
   ordinal: number
@@ -208,13 +202,6 @@ const validateArtifactExecutionInputs = (
   }
 }
 
-const validRecoveryFilename = (value: string): boolean =>
-  value.length > 0 &&
-  value !== '.' &&
-  value !== '..' &&
-  !value.includes('/') &&
-  !value.includes('\\')
-
 class ArtifactProvenanceRepository {
   private readonly compatibilityRepository: ArtifactRepository
   private readonly createId: () => string
@@ -224,6 +211,7 @@ class ArtifactProvenanceRepository {
   private readonly messageFinalizer: ArtifactProvenanceMessageFinalizer
   private readonly producerCapture: ArtifactProvenanceProducerCapture
   private readonly stagingRecovery: ArtifactProvenanceStagingRecovery
+  private readonly unindexedRecovery: ArtifactProvenanceUnindexedRecovery
   private readonly versionWriter: ArtifactProvenanceVersionWriter
 
   constructor(private readonly options: ArtifactProvenanceRepositoryOptions) {
@@ -259,6 +247,12 @@ class ArtifactProvenanceRepository {
       durability: this.durability,
       projectVersionFile: (version, projectId, appSessionId) =>
         this.toArtifactVersionFile(version, projectId, appSessionId)
+    })
+    this.unindexedRecovery = new ArtifactProvenanceUnindexedRecovery({
+      storageRoot: options.storageRoot,
+      getClient: options.getClient,
+      compatibilityRepository: this.compatibilityRepository,
+      createId: this.createId
     })
     this.versionWriter = new ArtifactProvenanceVersionWriter({
       storageRoot: options.storageRoot,
@@ -490,28 +484,12 @@ class ArtifactProvenanceRepository {
       projectId,
       options?.projectReconciliation
     )
-    const provenanceRoot = resolveStorageKey(
-      this.options.storageRoot,
-      storageKey('artifacts', projectId, appSessionId, '.provenance')
-    )
     const result: ArtifactStorageReconciliationResult = {
       recoveredVersionIds: [],
       quarantinedVersionIds: [],
       recoveredMessageArtifacts: []
     }
-    const lineageEntries = await readdir(provenanceRoot, { withFileTypes: true }).catch(
-      (error: unknown) => {
-        if (
-          typeof error === 'object' &&
-          error !== null &&
-          'code' in error &&
-          (error as { code?: unknown }).code === 'ENOENT'
-        ) {
-          return []
-        }
-        throw error
-      }
-    )
+    const unindexedSnapshot = await this.unindexedRecovery.prepareSession(projectId, appSessionId)
     const stagingResult = await this.stagingRecovery.reconcileSession(
       projectId,
       appSessionId,
@@ -519,7 +497,6 @@ class ArtifactProvenanceRepository {
     )
     result.recoveredVersionIds.push(...stagingResult.recoveredVersionIds)
     result.quarantinedVersionIds.push(...stagingResult.quarantinedVersionIds)
-    const client = await this.options.getClient()
     const finalizationResult = await this.finalizationRecovery.reconcileSession(
       projectId,
       appSessionId,
@@ -529,424 +506,10 @@ class ArtifactProvenanceRepository {
     result.recoveredVersionIds.push(...finalizationResult.recoveredVersionIds)
     result.recoveredMessageArtifacts.push(...finalizationResult.recoveredMessageArtifacts)
 
-    for (const lineageEntry of lineageEntries) {
-      if (
-        !lineageEntry.isDirectory() ||
-        lineageEntry.name === '.staging' ||
-        lineageEntry.name === '.quarantine' ||
-        !SAFE_SEGMENT_PATTERN.test(lineageEntry.name)
-      ) {
-        continue
-      }
-      const versionsRoot = join(provenanceRoot, lineageEntry.name, 'versions')
-      const versionEntries = await readdir(versionsRoot, { withFileTypes: true }).catch(
-        (error: unknown) => {
-          if (
-            typeof error === 'object' &&
-            error !== null &&
-            'code' in error &&
-            (error as { code?: unknown }).code === 'ENOENT'
-          ) {
-            return []
-          }
-          throw error
-        }
-      )
-      for (const versionEntry of versionEntries) {
-        if (!versionEntry.isDirectory() || !SAFE_SEGMENT_PATTERN.test(versionEntry.name)) continue
-        const existing = await client.artifactVersion.findUnique({
-          where: { id: versionEntry.name },
-          select: { id: true }
-        })
-        if (existing) continue
-
-        const versionDirectory = join(versionsRoot, versionEntry.name)
-        try {
-          await this.recoverUnindexedFinalizedVersion({
-            projectId,
-            appSessionId,
-            artifactId: lineageEntry.name,
-            versionId: versionEntry.name,
-            versionDirectory
-          })
-          result.recoveredVersionIds.push(versionEntry.name)
-        } catch (error) {
-          // Do not turn an unreadable sidecar elsewhere in the Project into evidence that this
-          // immutable Version is unowned. An incomplete scan is retryable, not quarantine proof.
-          if (error instanceof ArtifactCompatibilityScanIncompleteError) continue
-          // A concurrent reconciler may have inserted the same immutable row. Re-check before moving
-          // anything; only a still-unowned directory is eligible for quarantine.
-          const wonByAnotherWriter = await client.artifactVersion.findUnique({
-            where: { id: versionEntry.name },
-            select: { id: true }
-          })
-          if (wonByAnotherWriter) continue
-          const quarantineDirectory = join(
-            provenanceRoot,
-            '.quarantine',
-            'recovered-unlinked',
-            lineageEntry.name,
-            `${versionEntry.name}-${this.createId()}`
-          )
-          await mkdir(dirname(quarantineDirectory), { recursive: true })
-          await rename(versionDirectory, quarantineDirectory)
-          result.quarantinedVersionIds.push(versionEntry.name)
-        }
-      }
-    }
+    const unindexedResult = await this.unindexedRecovery.reconcileSession(unindexedSnapshot)
+    result.recoveredVersionIds.push(...unindexedResult.recoveredVersionIds)
+    result.quarantinedVersionIds.push(...unindexedResult.quarantinedVersionIds)
     return result
-  }
-
-  private async recoverUnindexedFinalizedVersion(input: {
-    projectId: string
-    appSessionId: string
-    artifactId: string
-    versionId: string
-    versionDirectory: string
-  }): Promise<void> {
-    const evidenceJson = await readFile(join(input.versionDirectory, 'evidence.json'), 'utf8')
-    const evidence = recordValue(JSON.parse(evidenceJson))
-    if (!evidence || canonicalJson(evidence as CanonicalJson) !== evidenceJson) {
-      throw new Error('Recovered Artifact evidence is not canonical.')
-    }
-    const filename = stringValue(evidence.filename)
-    const versionNumber = numberValue(evidence.version_number)
-    const sizeBytes = numberValue(evidence.size_bytes)
-    const checksum = stringValue(evidence.checksum)
-    const createdAtValue = stringValue(evidence.created_at)
-    const conversation = recordValue(evidence.conversation)
-    if (
-      evidence.schema_version !== 1 ||
-      evidence.project_id !== input.projectId ||
-      evidence.app_session_id !== input.appSessionId ||
-      evidence.artifact_id !== input.artifactId ||
-      evidence.version_id !== input.versionId ||
-      !filename ||
-      !validRecoveryFilename(filename) ||
-      !Number.isInteger(versionNumber) ||
-      (versionNumber ?? 0) < 1 ||
-      !Number.isSafeInteger(sizeBytes) ||
-      (sizeBytes ?? -1) < 0 ||
-      !checksum ||
-      !SHA256_PATTERN.test(checksum) ||
-      !createdAtValue ||
-      Number.isNaN(Date.parse(createdAtValue)) ||
-      !conversation
-    ) {
-      throw new Error('Recovered Artifact evidence identity is invalid.')
-    }
-    const rootFrameId = assertSafeSegment(
-      stringValue(conversation.root_frame_id) ?? '',
-      'root frame id'
-    )
-    const agentFrameId = assertSafeSegment(
-      stringValue(conversation.agent_frame_id) ?? '',
-      'agent frame id'
-    )
-    const messageBranchId = assertSafeSegment(
-      stringValue(conversation.message_branch_id) ?? '',
-      'message branch id'
-    )
-    const runtimeSegmentId = assertSafeSegment(
-      stringValue(conversation.runtime_segment_id) ?? '',
-      'runtime segment id'
-    )
-    const promptMessageId = assertSafeSegment(
-      stringValue(conversation.prompt_message_id) ?? '',
-      'prompt message id'
-    )
-    const content = await readFile(join(input.versionDirectory, 'content'))
-    if (content.byteLength !== sizeBytes || sha256(content) !== checksum) {
-      throw new Error('Recovered Artifact content is corrupt.')
-    }
-
-    const producer = recordValue(evidence.producer)
-    const executionStatus = recordValue(evidence.execution_status)
-    let executionSnapshotJson: string | undefined
-    let executionSnapshotChecksum: string | undefined
-    let producerRunId: string | undefined
-    let producerRunIndex: number | undefined
-    let notebookSessionId: string | undefined
-    if (executionStatus?.state === 'available') {
-      executionSnapshotJson = await readFile(join(input.versionDirectory, 'execution.json'), 'utf8')
-      executionSnapshotChecksum = stringValue(evidence.execution_snapshot_checksum)
-      if (
-        !executionSnapshotChecksum ||
-        sha256(executionSnapshotJson) !== executionSnapshotChecksum
-      ) {
-        throw new Error('Recovered Artifact execution snapshot is corrupt.')
-      }
-      const execution = recordValue(JSON.parse(executionSnapshotJson))
-      producerRunId = stringValue(producer?.producer_run_id)
-      producerRunIndex = numberValue(producer?.run_index)
-      notebookSessionId = stringValue(producer?.notebook_session_id)
-      if (
-        !execution ||
-        execution.schemaVersion !== 2 ||
-        execution.producerRunId !== producerRunId ||
-        execution.producerRunIndex !== producerRunIndex
-      ) {
-        throw new Error('Recovered Artifact producer binding is invalid.')
-      }
-    }
-
-    const snapshot = await this.findOwningMessageSnapshot({
-      projectId: input.projectId,
-      appSessionId: input.appSessionId,
-      versionId: input.versionId,
-      rootFrameId,
-      agentFrameId,
-      messageBranchId
-    })
-    const pendingRoute = await this.compatibilityRepository.findPendingVersionRouting({
-      projectName: input.projectId,
-      artifactId: input.artifactId,
-      versionId: input.versionId
-    })
-    if (snapshot && pendingRoute) {
-      throw new Error('Recovered Artifact has conflicting finalized and pending ownership proofs.')
-    }
-    if (!snapshot && !pendingRoute) {
-      throw new Error('Recovered Artifact has no exact lifecycle ownership proof.')
-    }
-    if (
-      pendingRoute &&
-      (pendingRoute.versionNumber !== versionNumber ||
-        normalizeFilename(pendingRoute.filename) !== normalizeFilename(filename) ||
-        pendingRoute.checksum !== checksum ||
-        (pendingRoute.mimeType ?? undefined) !== (stringValue(evidence.content_type) ?? undefined))
-    ) {
-      throw new Error('Recovered Artifact pending routing does not match immutable evidence.')
-    }
-
-    const rawInputs = Array.isArray(evidence.inputs) ? evidence.inputs : []
-    const inputs = rawInputs.map((value, index) => {
-      const item = recordValue(value)
-      const sourceKind = stringValue(item?.source_kind)
-      const inputFileVersionId = stringValue(item?.input_file_version_id)
-      const sourceFileId = stringValue(item?.source_file_id)
-      const sourceProjectId = stringValue(item?.source_project_id)
-      const sourceSessionId = stringValue(item?.source_session_id)
-      const inputFilename = stringValue(item?.filename)
-      const inputChecksum = stringValue(item?.checksum)
-      const storageKeyValue = stringValue(item?.storage_key)
-      const inputSize = numberValue(item?.size_bytes)
-      if (
-        item?.ordinal !== index ||
-        (sourceKind !== 'artifact-version' && sourceKind !== 'upload-version') ||
-        !inputFileVersionId ||
-        !sourceFileId ||
-        !sourceProjectId ||
-        !sourceSessionId ||
-        !inputFilename ||
-        !inputChecksum ||
-        !storageKeyValue ||
-        !Number.isSafeInteger(inputSize)
-      ) {
-        throw new Error('Recovered Artifact input evidence is invalid.')
-      }
-      return {
-        id: this.createId(),
-        ordinal: index,
-        inputFileVersionId,
-        sourceKind,
-        sourceFileId,
-        sourceArtifactVersionId: sourceKind === 'artifact-version' ? inputFileVersionId : undefined,
-        sourceUploadVersionId: sourceKind === 'upload-version' ? inputFileVersionId : undefined,
-        sourceVersionNumber: numberValue(item.source_version_number),
-        sourceCreatedAt: stringValue(item.source_created_at)
-          ? new Date(stringValue(item.source_created_at)!)
-          : undefined,
-        sourceProjectId,
-        sourceSessionId,
-        filename: inputFilename,
-        contentType: stringValue(item.content_type),
-        sizeBytes: BigInt(inputSize!),
-        checksum: inputChecksum,
-        storageKey: storageKeyValue,
-        strongestAssociation: stringValue(item.strongest_association) ?? 'captured-version'
-      }
-    })
-
-    const client = await this.options.getClient()
-    await client.$transaction(async (transaction) => {
-      await transaction.fileOriginSession.upsert({
-        where: {
-          projectId_sessionId: { projectId: input.projectId, sessionId: input.appSessionId }
-        },
-        create: { projectId: input.projectId, sessionId: input.appSessionId },
-        update: {}
-      })
-      const normalizedFilename = normalizeFilename(filename)
-      const lineageByName = await transaction.artifactLineage.findUnique({
-        where: {
-          projectId_sessionId_normalizedFilename: {
-            projectId: input.projectId,
-            sessionId: input.appSessionId,
-            normalizedFilename
-          }
-        }
-      })
-      if (lineageByName && lineageByName.id !== input.artifactId) {
-        throw new Error('Recovered Artifact lineage conflicts with an existing filename identity.')
-      }
-      const lineageById = await transaction.artifactLineage.findUnique({
-        where: { id: input.artifactId }
-      })
-      if (
-        lineageById &&
-        (lineageById.projectId !== input.projectId ||
-          lineageById.sessionId !== input.appSessionId ||
-          lineageById.normalizedFilename !== normalizedFilename)
-      ) {
-        throw new Error('Recovered Artifact lineage ownership conflicts with SQLite.')
-      }
-      if (!lineageById) {
-        await transaction.artifactLineage.create({
-          data: {
-            id: input.artifactId,
-            projectId: input.projectId,
-            sessionId: input.appSessionId,
-            normalizedFilename,
-            filename
-          }
-        })
-      }
-      await transaction.artifactVersion.create({
-        data: {
-          id: input.versionId,
-          artifactId: input.artifactId,
-          versionNumber: versionNumber!,
-          filename,
-          artifactRunId: pendingRoute?.artifactRunId ?? `recovered-${input.versionId}`,
-          rootFrameId,
-          agentFrameId,
-          messageBranchId,
-          runtimeSegmentId,
-          promptMessageId,
-          notebookSessionId,
-          producerRunId,
-          producerRunIndex,
-          messageId: snapshot?.terminalMessageId,
-          messageSnapshotId: snapshot?.id,
-          state: snapshot ? 'finalized' : 'pending',
-          contentStorageKey: storageKey(
-            'artifacts',
-            input.projectId,
-            input.appSessionId,
-            '.provenance',
-            input.artifactId,
-            'versions',
-            input.versionId,
-            'content'
-          ),
-          evidenceStorageKey: storageKey(
-            'artifacts',
-            input.projectId,
-            input.appSessionId,
-            '.provenance',
-            input.artifactId,
-            'versions',
-            input.versionId,
-            'evidence.json'
-          ),
-          contentType: stringValue(evidence.content_type),
-          sizeBytes: BigInt(sizeBytes!),
-          checksum,
-          evidenceJson,
-          evidenceChecksum: sha256(evidenceJson),
-          executionSnapshotJson,
-          executionSnapshotChecksum,
-          executionSnapshotStorageKey: executionSnapshotJson
-            ? storageKey(
-                'artifacts',
-                input.projectId,
-                input.appSessionId,
-                '.provenance',
-                input.artifactId,
-                'versions',
-                input.versionId,
-                'execution.json'
-              )
-            : undefined,
-          executionSnapshotSchemaVersion: executionSnapshotJson ? 2 : undefined,
-          ...(inputs.length > 0 ? { inputs: { create: inputs } } : {}),
-          createdAt: new Date(createdAtValue)
-        }
-      })
-    })
-  }
-
-  private async findOwningMessageSnapshot(input: {
-    projectId: string
-    appSessionId: string
-    versionId: string
-    rootFrameId: string
-    agentFrameId: string
-    messageBranchId: string
-  }): Promise<{ id: string; terminalMessageId: string } | undefined> {
-    const client = await this.options.getClient()
-    const candidates = await client.artifactMessageSnapshot.findMany({
-      where: {
-        projectId: input.projectId,
-        sessionId: input.appSessionId,
-        rootFrameId: input.rootFrameId,
-        agentFrameId: input.agentFrameId,
-        messageBranchId: input.messageBranchId,
-        state: 'ready'
-      }
-    })
-    const matches: Array<{ id: string; terminalMessageId: string }> = []
-    for (const candidate of candidates) {
-      try {
-        const serialized = await readFile(
-          resolveStorageKey(this.options.storageRoot, candidate.storageKey),
-          'utf8'
-        )
-        if (!candidate.checksum || sha256(serialized) !== candidate.checksum) continue
-        const payload = recordValue(JSON.parse(serialized))
-        const messages = Array.isArray(payload?.messages) ? payload.messages : []
-        const messageRecords = messages.map(recordValue)
-        const messageIds = new Set(messageRecords.map((message) => stringValue(message?.id)))
-        const hasCompleteParentChain = messageRecords.every((message, index) => {
-          if (!message || !stringValue(message.id)) return false
-          const parentMessageId = stringValue(message.parentMessageId)
-          return index === 0
-            ? parentMessageId === undefined
-            : parentMessageId === stringValue(messageRecords[index - 1]?.id)
-        })
-        const terminal = recordValue(messages.at(-1))
-        const artifacts = Array.isArray(terminal?.artifacts) ? terminal.artifacts : []
-        const parts = Array.isArray(terminal?.parts) ? terminal.parts : []
-        const ownsVersion =
-          artifacts.some((artifact) => recordValue(artifact)?.versionId === input.versionId) ||
-          parts.some(
-            (part) =>
-              recordValue(part)?.type === 'artifact' &&
-              recordValue(part)?.versionId === input.versionId
-          )
-        if (
-          (payload?.schemaVersion === 2 || payload?.schemaVersion === 3) &&
-          payload?.snapshotId === candidate.id &&
-          payload.rootFrameId === input.rootFrameId &&
-          payload.agentFrameId === input.agentFrameId &&
-          payload.messageBranchId === input.messageBranchId &&
-          payload.terminalMessageId === candidate.terminalMessageId &&
-          messages.length === candidate.messageCount &&
-          messageIds.size === messages.length &&
-          hasCompleteParentChain &&
-          (payload.schemaVersion !== 3 ||
-            (Array.isArray(payload.activities) && Array.isArray(payload.activityGroups))) &&
-          terminal?.id === candidate.terminalMessageId &&
-          ownsVersion
-        ) {
-          matches.push({ id: candidate.id, terminalMessageId: candidate.terminalMessageId })
-        }
-      } catch {
-        // A corrupt candidate cannot prove ownership; another valid snapshot may still do so.
-      }
-    }
-    return matches.length === 1 ? matches[0] : undefined
   }
 
   async getLineage(
